move crd check in Initialize
